# Conflicts:
#	src/SFA.DAS.Payments.Application/Data/Configurations/PaymentModelConfiguration.cs
#	src/SFA.DAS.Payments.Audit.Application/Data/PeriodisedPaymentsEventModelDataTable.cs
#	src/SFA.DAS.Payments.Audit.Application/Mapping/EarningEvents/EarningEventProfile.cs
#	src/SFA.DAS.Payments.Audit.Model/PeriodisedPaymentsEventModel.cs
#	src/SFA.DAS.Payments.Messages.Core/Events/PeriodisedPaymentEvent.cs
#	src/SFA.DAS.Payments.Model.Core/Entities/PaymentModel.cs
#	src/SFA.DAS.Payments.ProviderPayments.Application.UnitTests/Mapping/FundingSourceToPaymentEventMappingTests.cs
#	src/SFA.DAS.Payments.ProviderPayments.Application/Data/ProviderPaymentDataTable.cs
#	src/SFA.DAS.Payments.ProviderPayments.Application/Mapping/ProviderPaymentsProfile.cs
#	src/SFA.DAS.Payments.RequiredPayments.Application/Mapping/RequiredPaymentsProfile.cs
#	src/SFA.DAS.Payments.RequiredPayments.Model/Entities/PaymentHistoryEntity.cs

+++ src/SFA.DAS.Payments.Model.Core/LearningAim.cs
@@ -15,7 +15,7 @@ namespace SFA.DAS.Payments.Model.Core
         public int PathwayCode { get; set; }
 
         public string FundingLineType { get; set; }
-        public int SequenceNumber { get; set; }
+        public long SequenceNumber { get; set; }
 
         public DateTime StartDate { get; set; }
